feat: social sign-up
- make org sign-up details a component
